Add webpack to make it SPA
