+++ package.json
@@ -32,7 +32,8 @@
   "scripts": {
     "website": "npm run main && npm run build",
     "main": "ts-node src/watch-list.ts",
-    "build": "ts-node src/index.tsx",
+    "build": "webpack --mode production",
+    "start": "webpack serve --mode development",
     "clean": "rimraf lib/ module/",
     "prepare": "git config --local core.hooksPath .githook",
     "prepublishOnly": "npm run clean && npm run build",
@@ -57,13 +58,18 @@
     "@types/node": "^14.14.20",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
+    "html-webpack-plugin": "^5.0.0-beta.4",
     "lint-staged": "^10.5.3",
     "mocha": "^8.2.1",
     "prettier": "^2.2.1",
     "rimraf": "^3.0.2",
+    "ts-loader": "^8.0.14",
     "ts-node": "^9.1.1",
     "ts-node-test-register": "^9.0.1",
-    "typescript": "^4.1.3"
+    "typescript": "^4.1.3",
+    "webpack": "^5.14.0",
+    "webpack-cli": "^4.3.1",
+    "webpack-dev-server": "^3.11.2"
   },
   "publishConfig": {
     "access": "public"
