+++ src/App.css
@@ -1,0 +1,11 @@
+.Item-Main {
+    display: flex;
+}
+
+.Item-Tag {
+    border-radius: 12px;
+    padding: 4px 8px;
+    background: var(--nc-bg-1);
+    color: var(--nc-tx-2);
+    border: 1px solid var(--nc-lk-1);
+}
